hide popover by installing click handler on map
This is a hack because a click within the popover can't be stopped
from bubbling to the map. We can't use stopPropagation, because React
handlers inside the popover listen for events *on the body*. They
won't fire if we try to stopPropagation when clicking inside the
popover.

But that means the click goes to the map, which means it will try to
close the popover.

You can see my disgusting solution to this problem